requirements.txt
greenlet==1.1.2
lxml==4.6.3
pony==0.7.14
psycopg2-binary==2.9.1
PyMySQL==1.0.2
pytz==2021.3
spyne==2.11.0
SQLAlchemy==0.9.4
suds-community==1.0.0
